Added new notebooks

+++ experiments/variability/train_htms2.py
@@ -186,34 +186,36 @@ def train_automatic(path_corpus: str,
             for iter_ in range(iters):
                 iter_ += start
                 for exp_tpc in range(ntopics_root):
-                    for tr_tpc in num_topics_sub:
+                    #if exp_tpc >= 11:
+                    if exp_tpc >= 0:
+                        for tr_tpc in num_topics_sub:
 
-                        training_params["ntopics"] = tr_tpc
+                            training_params["ntopics"] = tr_tpc
 
-                        name = f"submodel_{version}_from_topic_{str(exp_tpc)}_train_with_{str(tr_tpc)}_iter_{iter_}_{DT.datetime.now().strftime('%Y%m%d')}"
+                            name = f"submodel_{version}_from_topic_{str(exp_tpc)}_train_with_{str(tr_tpc)}_iter_{iter_}_{DT.datetime.now().strftime('%Y%m%d')}"
 
-                        submodel_path = tm_wrapper.train_htm_submodel(
-                            version=version,
-                            father_model_path=model_path_parent,
-                            name=name,
-                            trainer=trainer,
-                            training_params=training_params,
-                            expansion_topic=exp_tpc,
-                            thr=None
-                        )
+                            submodel_path = tm_wrapper.train_htm_submodel(
+                                version=version,
+                                father_model_path=model_path_parent,
+                                name=name,
+                                trainer=trainer,
+                                training_params=training_params,
+                                expansion_topic=exp_tpc,
+                                thr=None
+                            )
 
-                        # Recalculate coherence vs ref corpus
-                        logger.info(
-                            f"-- -- Recalculating coherence vs ref corpus")
-                        for metric in ["c_v", "c_npmi"]:
-                            tm_wrapper.calculate_cohr_vs_ref(
-                                model_path=submodel_path,
-                                corpus_val=pathlib.Path(path_ref_corpus),
-                                type=metric)
-                        logger.info(
-                            f"-- -- Calculating RBO and TD")
-                        tm_wrapper.calculate_rbo(submodel_path)
-                        tm_wrapper.calculate_td(submodel_path)
+                            # Recalculate coherence vs ref corpus
+                            logger.info(
+                                f"-- -- Recalculating coherence vs ref corpus")
+                            for metric in ["c_v", "c_npmi"]:
+                                tm_wrapper.calculate_cohr_vs_ref(
+                                    model_path=submodel_path,
+                                    corpus_val=pathlib.Path(path_ref_corpus),
+                                    type=metric)
+                            logger.info(
+                                f"-- -- Calculating RBO and TD")
+                            tm_wrapper.calculate_rbo(submodel_path)
+                            tm_wrapper.calculate_td(submodel_path)
             # ----> END HTM-WS
 
         if htm_ds:
@@ -227,43 +229,45 @@ def train_automatic(path_corpus: str,
             for iter_ in range(iters):
                 iter_ += start
                 for exp_tpc in range(ntopics_root):
-                    for tr_tpc in num_topics_sub:
-                        for thr in np.arange(0.1, 1, 0.1):
-                            logger.info(f"-- -- Generating submodel with HTM-DS thr {thr} and {tr_tpc} topics")
-                            try:
-                                thr_f = "{:.1f}".format(thr)
-                                name = f"submodel_{version}_thr_{thr_f}_from_topic_{str(exp_tpc)}_train_with_{str(tr_tpc)}_iter_{iter_}_{DT.datetime.now().strftime('%Y%m%d')}"
-                               
-                                submodel_path = tm_wrapper.train_htm_submodel(
-                                    version=version,
-                                    father_model_path=pathlib.Path(model_path_parent),
-                                    name=name,
-                                    trainer=trainer,
-                                    training_params=training_params,
-                                    expansion_topic=exp_tpc,
-                                    thr=thr
-                                )
+                    #if exp_tpc >= 11:
+                    if exp_tpc >= 0:
+                        for tr_tpc in num_topics_sub:
+                            for thr in np.arange(0.1, 1, 0.1):
+                                logger.info(f"-- -- Generating submodel with HTM-DS thr {thr} and {tr_tpc} topics")
+                                try:
+                                    thr_f = "{:.1f}".format(thr)
+                                    name = f"submodel_{version}_thr_{thr_f}_from_topic_{str(exp_tpc)}_train_with_{str(tr_tpc)}_iter_{iter_}_{DT.datetime.now().strftime('%Y%m%d')}"
+                                
+                                    submodel_path = tm_wrapper.train_htm_submodel(
+                                        version=version,
+                                        father_model_path=pathlib.Path(model_path_parent),
+                                        name=name,
+                                        trainer=trainer,
+                                        training_params=training_params,
+                                        expansion_topic=exp_tpc,
+                                        thr=thr
+                                    )
 
-                                # Recalculate coherence vs ref corpus
-                                logger.info(
-                                    f"-- -- Recalculating coherence vs ref corpus")
-                                for metric in ["c_v", "c_npmi"]:
-                                    tm_wrapper.calculate_cohr_vs_ref(
-                                        model_path=submodel_path,
-                                        corpus_val=pathlib.Path(
-                                            path_ref_corpus),
-                                        type=metric)
-                                logger.info(
-                                    f"Calculating RBO and TD")
-                                tm_wrapper.calculate_rbo(submodel_path)
-                                tm_wrapper.calculate_td(submodel_path)
-                            except Exception as error:
-                                logger.error(f"Error with thr {thr}")
-                                logger.error(error)
-                                this_submodel_path = pathlib.Path(
-                                    model_path_parent).joinpath(name)
-                                if this_submodel_path.is_dir():
-                                    shutil.rmtree(this_submodel_path)
+                                    # Recalculate coherence vs ref corpus
+                                    logger.info(
+                                        f"-- -- Recalculating coherence vs ref corpus")
+                                    for metric in ["c_v", "c_npmi"]:
+                                        tm_wrapper.calculate_cohr_vs_ref(
+                                            model_path=submodel_path,
+                                            corpus_val=pathlib.Path(
+                                                path_ref_corpus),
+                                            type=metric)
+                                    logger.info(
+                                        f"Calculating RBO and TD")
+                                    tm_wrapper.calculate_rbo(submodel_path)
+                                    tm_wrapper.calculate_td(submodel_path)
+                                except Exception as error:
+                                    logger.error(f"Error with thr {thr}")
+                                    logger.error(error)
+                                    this_submodel_path = pathlib.Path(
+                                        model_path_parent).joinpath(name)
+                                    if this_submodel_path.is_dir():
+                                        shutil.rmtree(this_submodel_path)
 
 
 def main():
